Only load jinja2 when template format is loaded.
